# docs/swagger.yaml
definitions:
  models.CreateSubsRequest:
    properties:
      end_date:
        example: 07-2025 or null (it will be infinity)
        type: string
      price:
        example: 299
        type: integer
      service_name:
        example: Spotify
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  models.ErrorResponse:
    properties:
      msg:
        type: string
    type: object
  models.GetSubsResponse:
    properties:
      end_date:
        example: 07-2025 or 12-9999(infinity)
        type: string
      price:
        example: 299
        type: integer
      service_name:
        example: Spotify
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  models.GoodResponse:
    properties:
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/create-subs:
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку по данным из тела запроса
      parameters:
      - description: Данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GoodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /api/v1/delete-subs/{user_id}:
    delete:
      description: Удаляет подписку, связанную с указанным пользователем
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GoodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Удалить подписку по user_id
      tags:
      - subscriptions
  /api/v1/get-subs:
    get:
      description: Возвращает список всех подписок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetSubsResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить все подписки
      tags:
      - subscriptions
  /api/v1/get-subs/{user_id}:
    get:
      description: Возвращает информацию о подписке по ID пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetSubsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить подписку по user_id
      tags:
      - subscriptions
  /api/v1/update-subs/{user_id}:
    put:
      consumes:
      - application/json
      description: Обновляет данные подписки по данным из тела запроса
      parameters:
      - description: Обновлённые данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetSubsResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
swagger: "2.0"